# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Uploy API
  version: 0.2.0
  description: Uploy deployment platform API
paths:
...
  /api/servers/{id}/proxy-logs:
    get:
      operationId: getProxyLogs
      summary: Stream the reverse proxy's logs via SSE
      description: |
        Returns a Server-Sent Events stream of Traefik's output on this server —
        its own startup and certificate messages, and one access-log line per
        request it routes. OpenAPI does not natively model SSE, so this endpoint
        is documented for reference only.

        This is per server rather than per service: one proxy answers for every
        service on the machine, so the log is one stream and the service a line
        belongs to is named in the line itself.

        **SSE events (not modeled by this schema):**
        - Default message event: JSON object with "output" and "type" ("stdout" or "stderr")
        - `done` event: the proxy's stream ended
        - `stream-error` event: JSON object with "message" field
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          required: false
          description: >-
            Only replay history from this far back. Omitted, the stream starts
            from the last 200 lines however old they are.
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
      responses:
        "200":
          description: SSE log stream
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: "`since` is not one of the listed values"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/projects:
    post:
      operationId: createProject
      summary: Create a new project
      security:
        - cookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listProjects
      summary: List projects in workspace
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/projects/from-image:
    post:
      operationId: createProjectFromImage
      summary: Create a project, default environment, and Docker image service atomically
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectFromImageRequest"
      responses:
        "201":
          description: Project, environment, and service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectFromImageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Container name already in use on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/projects/{id}:
    get:
      operationId: getProject
      summary: Get project details
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      summary: Update project
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      summary: Delete a project
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Project deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
          description: >
            Machine-readable failure stage. Present on SSH probe failures so a
            client can offer the right remedy: the three stages need entirely
            different fixes, and the human message alone cannot be branched on.
          enum:
            - unreachable
            - key_rejected
            - key_invalid
            - session_failed
            - docker_missing
...
    ProjectResponse:
      type: object
      required:
        - id
        - name
        - workspace_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        workspace_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      description: |
        Request body for creating a project. When `name` is omitted, blank, or
        whitespace-only, the backend generates a unique Railway-style
        `adjective-noun` name scoped to the workspace.
      properties:
        name:
          type: string
    UpdateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    CreateProjectFromImageRequest:
      type: object
      required:
        - server_id
        - image
        - container_port
      properties:
        server_id:
          type: string
        image:
          type: string
          description: Docker image reference, e.g. `nginx:latest` or `ghcr.io/owner/repo:tag`.
        container_port:
          type: integer
          description: Container port the service listens on.
        host_port:
          type: integer
          description: >-
            Host port the service is published on. Omit to keep the service
            internal — reachable by other services on the uploy network and by
            nothing outside the machine, which is what a database wants. 80 and
            443 are rejected: they belong to the Uploy proxy. Ignored while the
            service has a domain, since that traffic goes through the proxy.
    CreateProjectFromImageResponse:
      type: object
      required:
        - project
        - environment
        - service
      properties:
        project:
          $ref: "#/components/schemas/ProjectResponse"
        environment:
          $ref: "#/components/schemas/EnvironmentResponse"
        service:
          $ref: "#/components/schemas/ServiceResponse"
    EnvironmentResponse:
      type: object
      required:
        - id
        - name
        - project_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        project_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ServiceResponse:
      type: object
      required:
        - id
        - name
        - image
        - container_name
        - container_port
        - server_id
        - kind
        - project_id
        - environment_id
        - created_at
        - updated_at
        - pending_change_count
        - has_deployed
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
        container_name:
          type: string
        container_port:
          type: integer
          description: Container port the image listens on.
        host_port:
          type: integer
          description: >-
            Host port the service is published on. Omit to keep the service
            internal — reachable by other services on the uploy network and by
            nothing outside the machine, which is what a database wants. 80 and
            443 are rejected: they belong to the Uploy proxy. Ignored while the
            service has a domain, since that traffic goes through the proxy.
        server_id:
          type: string
        kind:
          type: string
          enum:
            - application
        project_id:
          type: string
        environment_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        pending_change_count:
          type: integer
          description: >
            How many changes are waiting to be deployed: the difference between
            the service's current configuration and the configuration its last
            successful deployment actually shipped, counted per field, per
            domain and per variable. Zero means the running container matches
            what the service says it is.

            A service with nothing to compare against — never deployed, or last
            deployed before Uploy recorded configurations — reports 1. It is
            pending, but there is no itemised answer to give.

            Derived per request, never stored.
        has_deployed:
          type: boolean
          description: >
            True once at least one deployment of this service has succeeded.
            Combined with has_pending_changes it distinguishes a service that
            will be created on its server from one that will be updated.
            Derived per request, never stored.
...